dspy-service/requirements.txt
# Core DSPy and Google dependencies (lightweight)
dspy-ai>=2.4.0
google-generativeai>=0.8.0

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Essential data processing (lightweight)
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Logging
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
